review-controller

+++ DTO/ReviewDTO.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int AppointmentId { get; set; }
         public string ClientUsername { get; set; }
-        public int ClientRating { get; set; }
+        public decimal ClientRating { get; set; }
         public string ClientComment { get; set; }
         public DateTime CreationDate { get; set; } = DateTime.Now;
     }
